moved resets to main, renamed homepage, moved pages to separate folder
