+++ src/css/AboutPage.css
@@ -1,5 +1,3 @@
-@import './style-variables.css';
-
 .about-page {
   /* align-items: center; */
   justify-content: center;
