+++ src/css/MovieCard.css
@@ -1,11 +1,3 @@
-/* ========================================================================== */
-/* ========================================================================== */
-/* Imports */
-@import "./style-variables.css";
-
-/* ========================================================================== */
-/* Main */
-
 #card-modal {
   background-color: var(--red);
 }
