+++ src/css/main.css
@@ -1,6 +1,15 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,500,700,900');
-@import './reset.css';
+@import url('https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900');
 @import './style-variables.css';
+
+/* =========================== RESETS =========================== */
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+/* =========================== MAIN =========================== */
 
 body {
   margin-top: 55px;
